perf(shaders): cut texture fetches + fix GL2.1 driver compat
- blur: 7-tap gaussian folded into 5 bilinear taps (~29% fewer fetches/pass)
- halo/pylon_glow: hoist center texel sample (was re-fetched up to 5x)
- doppleganger: compute rand() once instead of 3x
- downsample_13tap: drop dead TentFilter; perlin: drop unused color calc
- declare 'time' as float to match glUniform1f upload (int froze effects on strict/low-end GPUs)
- replace float[]() array ctors with plain consts; float literals in min/max/clamp

+++ src/main/resources/assets/npcdbc/shader/doppleganger.vert
@@ -1,12 +1,12 @@
 #version 120
 
 varying vec2 texcoord;
-uniform int time; // Passed in, see ShaderHelper.java
+uniform float time; // Passed in, see ShaderHelper.java (uploaded via glUniform1f)
 
 uniform float disfiguration; // Passed in via Callback
 
 float rand(vec2 co) {
-   return (fract(sin(dot(co.xy, vec2(12.9898,78.233))) * 43758.5453) - 0.5) * 2;
+   return (fract(sin(dot(co.xy, vec2(12.9898,78.233))) * 43758.5453) - 0.5) * 2.0;
 }
 
 void main() {
