+++ src/main/resources/assets/npcdbc/shader/doppleganger_bar.frag
@@ -1,7 +1,7 @@
 #version 120
 
 uniform sampler2D bgl_RenderedTexture;
-uniform int time; // Passed in, see ShaderHelper.java
+uniform float time; // Passed in, see ShaderHelper.java (uploaded via glUniform1f)
 
 uniform int startX; // Passed in, see BossBarHandler.java
 uniform int startY; // Passed in, see BossBarHandler.java
@@ -24,22 +24,22 @@ void main() {
     float s = sin((texcoord.y + 0.0375) * 100.0 + sin(texcoord.x * 90.0 - time / 2.0) * hpFract) * 0.5 + 0.5;
     s = 1.0 - s;
     if(s <= 0.98)
-        s = 0;
+        s = 0.0;
     s *= 0.4;
-    
-    r = min(1, r + s);
-    g = min(1, g + s);
-    b = min(1, b + s);
-    
+
+    r = min(1.0, r + s);
+    g = min(1.0, g + s);
+    b = min(1.0, b + s);
+
     if(hpFract <= 0.2) {
-        float flash = (sin(time * 2.0) - 0.8) * 5;
-        if(flash > 0) {
-            g = max(0, g - flash);
-            b = max(0, b - flash);
-        }   
+        float flash = (sin(time * 2.0) - 0.8) * 5.0;
+        if(flash > 0.0) {
+            g = max(0.0, g - flash);
+            b = max(0.0, b - flash);
+        }
     }
-    
-    if(grainIntensity > 0) {
+
+    if(grainIntensity > 0.0) {
         float gs = rand(vec2(texcoord.x + time, texcoord.y));
         r = r * (1 - grainIntensity) + gs * grainIntensity;
         g = g * (1 - grainIntensity) + gs * grainIntensity;
