+++ src/main/resources/assets/npcdbc/shader/gold.frag
@@ -1,7 +1,7 @@
 #version 120
 
 uniform sampler2D bgl_RenderedTexture;
-uniform int time; // Passed in, see ShaderHelper.java
+uniform float time; // Passed in, see ShaderHelper.java (uploaded via glUniform1f)
 
 void main() {
     vec2 texcoord = vec2(gl_TexCoord[0]);
